requirements.txt
openai
Pillow
customtkinter
